set the models/ layer for thr types
