+++ src/components/block/ExternalAuth/ExternalAuth.types.ts
@@ -1,0 +1,5 @@
+export interface ExternalAuthProps {
+    authTittle: string
+    src: string
+    alt: string
+}
