+++ src/components/block/Input/Input.types.ts
@@ -1,16 +1,16 @@
 import { Path } from "react-hook-form"
 import { RegisterOptions } from "react-hook-form"
-export interface InputField {
+
+export interface InputFieldType {
     username: string
-    email: string
+    email: string 
     password: string
     confirm_password: string
 }
 
 export interface InputProps {
     placeholder: string
-    field: Path<InputField>
     type: string
-    options: RegisterOptions<InputField>
+    field: Path<InputFieldType>
+    options: RegisterOptions<InputFieldType>
 }
-
