+++ src/components/block/NavBar/NavBar.tsx
@@ -1,7 +1,7 @@
 "use client"
 import React from "react";
 import { BiSolidMoon } from "react-icons/bi"
-import { NavBarType } from "@/entities/compoent-entities/header-entities";
+import { NavBarType } from "./NavBar.types";
 
 export const NavBar = ({firstItem,secondItem,thirdItem}: NavBarType) => {
     return(
