+++ src/components/block/NavBar/NavBar.types.ts
@@ -1,8 +1,3 @@
-export interface IconRouteType {
-    iconComponent: React.ReactElement
-    href: string
-}
-
 export interface NavBarType{
     firstItem : string;
     secondItem: string
