+++ src/components/sections/Header/Header.types.ts
@@ -1,0 +1,9 @@
+export interface HeaderProps {
+    barComponent?: React.ReactNode
+    iconList: IconRouteType[]
+}
+
+export interface IconRouteType {
+    iconComponent: React.ReactElement
+    href: string
+}
