+++ src/components/sections/Hero/Hero.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import Image from "next/image"
-import { HeroSectionType } from "@/entities/compoent-entities/hero-entities"
+import { HeroType } from "./Hero.types"
 
-export const Hero = ({titleHero,description,buttonContent}: HeroSectionType)=> {
+export const Hero = ({titleHero,description,buttonContent}: HeroType)=> {
     return(
         <section className=" flex md:justify-around justify-center items-center md:py-5 py-3 md:my-4 my-2">
             <div className=" flex flex-col md:w-1/3 w-9/12 gap-4 items-center my-6 mx-4 ">
