+++ src/components/sections/Hero/Hero.types.ts
@@ -1,4 +1,4 @@
-export interface HeroSectionType {
+export interface HeroType {
     titleHero: string
     description: string
     buttonContent: string
