--- src/entities/application-entities/user.ts
@@ -1,5 +1,0 @@
-export interface UserType {
-    name: string;
-    lastname: string;
-    photoUrl: string;
-}
