--- src/entities/compoent-entities/input-entities.ts
@@ -1,1 +1,0 @@
-
